Fix : fix issue of dbContext

+++ AuthService/Dtos/User/UserCodeDto.cs
@@ -1,4 +1,4 @@
 namespace AuthService.Dtos.User;
-public record UserCodeDto(int Id , string Code);
+public record UserCodeDto(long Id , string Code);
